+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to the Wazuh app for Splunk project will be documented in th
 - Configuration section ([#261](https://github.com/wazuh/wazuh-splunk/pull/261))
 - FIM section ([#255](https://github.com/wazuh/wazuh-splunk/pull/255)).
 - Settings section ([#265](https://github.com/wazuh/wazuh-splunk/pull/265)).
+- The `wz-table` directive now checks if a request is aborted ([]()).
+
 ### Fixes
 - Minor fixes.
 ## Wazuh v3.6.1 - Splunk Enterprise v7.1.3 - Splunk app v3.6.1-rev-16